fix race condition when resizing images - covers could be a duplicate of an existing image
